% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperest.R
\name{hyperest}
\alias{hyperest}
\title{Estimate treatment effects using hypothetical evaluations}
\usage{
hyperest(
  dat_wide,
  Y_name,
  W_name,
  H_names,
  H0_names,
  H1_names,
  X_names = NULL,
  ...
)
}
\arguments{
\item{dat_wide}{dataframe or equivalent, holding data on separate (assumed to be independent) observations as rows,
with column corresponding to different variables.}

\item{Y_name}{string, the name of the column of dat_wide that holds the outcome data.}

\item{W_name}{string, the name of the column of dat_wide that holds the treatment indicator. It is only used to estimate the ATT and ATC.}

\item{H_names}{string vector, the names of columns holding hypothetical evaluations for the realized treatment state.}

\item{H0_names}{string vector, the names of columns holding hypothetical evaluations in the absence of treatment.
IMPORTANT: order of variables must be the same as in H_names.}

\item{H1_names}{string vector, the names of columns holding hypothetical evaluations in the presence of treatment.
IMPORTANT: order of variables must be the same as in H_names.}

\item{X_names}{(optional) string vector, the names of columns holding fixed characteristics that do not vary by treatment state.}

\item{...}{(optional) parameters passed to vcovHC of the sandwich package,
determining how to calculate heteroskedasticity-robust standard errors for the linear regression in step 1.
If no parameters are supplied, the defaults of sandwich::vcovHC are used.}
}
\value{
list:
        \item{tau}{estimate of the average treatment effect}
        \item{tau_att}{estimate of the average treatment effect on the treated (ATT)}
        \item{tau_atc}{estimate of the average treatment effect on the control (ATC)}
        \item{se}{standard error of the ATE estimator}
        \item{se_att}{standard error of the ATT estimator}
        \item{se_atc}{standard error of the ATC estimator}
        \item{se_naive_only_s1}{standard error of the ATE estimator taking into account only the step 1 regression variance (holding fixed hypothetical evaluations in step 2)}
        \item{se_naive_only_s2}{standard error of the ATE estimator taking into account only the step 2 variance due to sampling of hypothetical evaluations (holding fixed the step 1 coefficient estimates)}
        \item{reg}{`lm` regression object of step 1}
        \item{tau_i}{predictions of individual-level treatment effects}
}
\description{
This function estimates treatment effects and standard errors using hypothetical evaluations.
}
\details{
Step 1 estimates a linear regression:
  \deqn{Y ~ H + X}
where the outcome variable Y is given in the input string Y_name,
the hypothetical evaluations for the realized treatment state are H given in the input string vector of column names H_names,
and optional fixed characteristics (not varying by treatment state) are given by the input string vector of column names X_names.

Step 2 estimates treatment effects as the difference in predictions for the treated and control state:
  \deqn{\tau = (H(1) - H(0)) \beta}
where H(1) are the hypothetical evaluations for the treated state given by the input string column names H1_names,
H(0) are the hypothetical evaluations for the control state given by the input column names H0_names,
and \eqn{\beta} are the estimated coefficients on H in the regression of step 1.

The function estimates the ATE, ATT, ATC, and their standard errors.
It also predicts individual level treatment effects, but no standard errors are provided
because theoretical properties are only available for average effects within groups (that grow with sample size).
}
\examples{
# simulate a data set with 500 observations, 3 hypothetical variables, 2 fixed characteristics
sims <- simulate_data(n=500, num_h=3, num_x=2)
# extract data
dat_wide <- sims$dat_wide
# estimate effects using hypothetical evaluations; match column names generated inside simulate_data
est_hyp <- hyperest(dat_wide,
                    Y_name="Y",
                    W_name="W",
                    H_names=paste0("H",1:3),
                    H0_names=paste0("H",1:3,"_0"),
                    H1_names=paste0("H",1:3,"_1"),
                    X_names=paste0("X",1:2))
# compare true (in-sample) effect, estimates using hypothetical evaluations, and difference in means
round(c(sims$tau,
        est_hyp$tau,
        mean(sims$dat_wide$Y[dat_wide$W==TRUE]) - mean(sims$dat_wide$Y[dat_wide$W==FALSE])),
      2)

# for additional verification, plot predicted treatment effects against individual-level effects
graphics::plot(est_hyp$tau_i,sims$tau_i)
reg <- stats::lm(tau ~ tau_hat, data = data.frame(tau = sims$tau_i, tau_hat = est_hyp$tau_i))
reg$coefficients
graphics::abline(reg)

}
\references{
B. Douglas Bernheim, Daniel Björkegren, Jeffrey Naecker, and Michael Pollmann. "Causal Inference from Hypothetical Evaluations." 2021.
}
