% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate a data set of hypothetical evaluations}
\usage{
simulate_data(n = 300, num_h = 5, num_x = 5, frac_coeff_0 = 0)
}
\arguments{
\item{n}{integer, sample size, must be large enough to allow regression on hypotheticals and fixed characteristics.}

\item{num_h}{integer, number of hypothetical evaluation variables, must be at least 1.}

\item{num_x}{integer, number of fixed characteristics, can be 0.}

\item{frac_coeff_0}{numeric, between 0 and 1, probability of each coefficient being 0.}
}
\value{
list:
         \item{dat_wide}{data.frame with `n` rows and `4+3*num_h+num_x` columns.
                           Column Y holds observed outcomes.
                           Column W holds realized treatment state.
                           Columns H1 through H<num_h> hold hypothetical evaluations corresponding to the realized treatment state.
                           Columns H1_0 through H<num_h>_0 hold hypothetical evaluations corresponding to the control state.
                           Columns H1_1 through H<num_h>_1 hold hypothetical evaluations corresponding to the treated state.
                           Columns X1 through X<num_x> hold fixed characteristics.
                           Columns Y1 and Y0 hold potential outcomes (under treatment and control) that are not typically available in real data sets}
         \item{dat_long}{data.frame with `2*n` rows and `2+num_h+num_x` columns.
                           The first `n` rows contain data under the control state of each observation,
                           the last `n` rows contain data under the treated state of each observation;
                           observations are in the same order in each half.
                           Column Y holds the outcome (if observed under the treatment state, otherwise NA).
                           Column W indicates the treatment state for which the row contains data.
                           Columns H1 through H<num_h> hold hypothetical evaluations for the treatment state of the row.
                           Columns X1 through X<num_x> hold fixed characteristics.}
         \item{tau_i}{vector of length `n`. Individual-level treatment effects (in the same order as the observations in `dat_wide` and `dat_long`.}
         \item{tau}{numeric. The true in-sample average treatment effect.}
}
\description{
Simulate a data set of hypothetical evaluations
}
\details{
This function simulates data that can be used to illustrate to illustrate the estimation methods included in hypeRest package.
The data generating process is **not** calibrated to resemble any real data set.
The data are simulated such that the simple difference in means estimator is (severely) biased.
}
\examples{
# simulate a data set with 500 observations, 3 hypothetical variables, 2 fixed characteristics
sims <- simulate_data(n=500, num_h=3, num_x=2)
# extract data
dat_wide <- sims$dat_wide
# estimate effects using hypothetical evaluations; match column names generated inside simulate_data
est_hyp <- hyperest(sims$dat_wide,
                    Y_name="Y",
                    W_name="W",
                    H_names=paste0("H",1:3),
                    H0_names=paste0("H",1:3,"_0"),
                    H1_names=paste0("H",1:3,"_1"),
                    X_names=paste0("X",1:2))
# compare true (in-sample) effect, estimates using hypothetical evaluations, and difference in means
round(c(sims$tau,
        est_hyp$tau,
        mean(dat_wide$Y[dat_wide$Y==TRUE]) - mean(dat_wide$Y[dat_wide$Y==FALSE])),
      2)

}
